reintroduce native threads for the updater as a setting and cleanup

+++ core/src/main/java/com/volmit/iris/engine/mantle/EngineMantle.java
@@ -249,7 +249,7 @@ public interface EngineMantle extends MatterGenerator {
         }
     }
 
-    default long getUnloadRegionCount() {
+    default int getUnloadRegionCount() {
         return getMantle().getUnloadRegionCount();
     }
